Removed more shapes
