+++ src/main/kotlin/com/manimdsl/runtime/VirtualMachine.kt
@@ -21,7 +21,6 @@ import com.manimdsl.runtime.datastructures.stack.StackValue
 import com.manimdsl.runtime.utility.getBoundaries
 import com.manimdsl.runtime.utility.wrapCode
 import com.manimdsl.runtime.utility.wrapString
-import com.manimdsl.shapes.SubtitleBlockShape
 import com.manimdsl.stylesheet.PositionProperties
 import com.manimdsl.stylesheet.Stylesheet
 import java.util.*
@@ -405,7 +404,7 @@ class VirtualMachine(
             }
             linearRepresentation.add(
                 UpdateSubtitle(
-                    subtitleBlockVariable.shape as SubtitleBlockShape,
+                    (subtitleBlockVariable as SubtitleBlock),
                     wrapString(text, 30),
                     runtime = animationSpeeds.first()
                 )
